VTrade_BE/requirements.txt
asgiref==3.8.1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
Django==5.1.1
django-allauth==65.0.2
django-phonenumber-field==8.0.0
google-api-core==2.20.0
google-api-python-client==2.147.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.65.0
httplib2==0.22.0
idna==3.10
phonenumbers==8.13.47
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.28.2
psycopg2-binary>=2.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyparsing==3.1.4
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
sqlparse==0.5.1
uritemplate==4.1.1
urllib3==2.2.3
